// symphony/opera/app/types.d.ts
export interface SolutionVersionSpec {
    displayName: string;
    components: ComponentSpec[];
    metadata: Record<string, string>;
}

export interface ComponentSpec {
    name: string;
    type: string;
    properties: Record<string, any>;
    metadata: Record<string, string>;
    parameters: Record<string, string>;
    routes: RouteSpec[];
    constraints: string;
    depedencies: string[];
    skills: string[];
    sidecars: SidecarSpec[];
}

export interface RouteSpec {
    route: string;
    type: string;
    properties: Record<string, string>;
    filters: filterSpec[];
}

export interface filterSpec {
    direction: string;
    type: string;
    parameters: Record<string, string>;
}

export interface SidecarSpec {
    name: string;
    type: string;
    properties: Record<string, any>;
}